chore: update zip function
- add overwrite option

+++ vlr/data/utils/tools.py
@@ -17,12 +17,12 @@ def clean_up(channel_name: str, dirs: list, overwrite: bool = False):
             os.makedirs(dir)
 
 
-def zip_dir(zip_dir: str):
+def zip_dir(zip_dir: str, overwrite: bool = False):
     """
     Zip directory.
     :param dir:         Path to directory.
     :param zip_path:    Path to zip file.
     """
-    if os.path.exists(zip_dir + ".zip"):
+    if overwrite and os.path.exists(zip_dir + ".zip"):
         os.remove(zip_dir + ".zip")
     shutil.make_archive(zip_dir, "zip", os.path.dirname(zip_dir), os.path.basename(zip_dir))
